Added init_database file, made some changes in Doc

+++ src/reval/controller.ts
@@ -5,64 +5,63 @@ import { Details, RevalRow, RevalTable, SubjectDetails } from '../interfaces/rev
 import dayjs from "dayjs";
 import { isAnyUndefined, responses } from "../services/common";
 
-// Revaluation Search
+// ANCHOR Revaluation Search
 
 export async function revalSearch(req: Request, res: Response) {
-  const rollNo: string = req.query.rollNo as string;
-  const exMonth: number = parseInt(req.query.exMonth as string);
-  const exYear: number = parseInt(req.query.exYear as string);
-
-  if (isAnyUndefined(rollNo, exMonth, exYear)) {
-    res.status(400).json(responses.NotAllParamsGiven);
-    return;
-  }
-  let subjects: Details = {};
-
-  let subjDetails: { subCodes: string[]; subNames: string[] }[] = [];
-  try {
-    const result = (await dbQuery(
-      `SELECT subCode, subName FROM printReval WHERE rollNo = '${rollNo}';`
-    )) as RevalRow[];
-    // Checking whether the std is already in the print table for registration otherwise fetching std details from the studentInfo table which are not paid
-    const query: string =
-      result.length > 0
-        ? `SELECT subCode, subName FROM printReval WHERE rollNo = ? AND acYear = ? AND sem = ?`
-        : `SELECT std.subCode, std.subName FROM studentInfo std LEFT JOIN paidReEvaluation paidStd ON std.subCode = paidStd.subCode AND std.rollNo = paidStd.rollNo WHERE std.rollNo = ? AND std.exMonth = ${exMonth} AND std.exYear = ${exYear} AND std.acYear = ? AND std.sem = ? AND paidStd.subCode IS NULL AND paidStd.rollNo IS NULL`;
-    let year: number = 1,
-      sem: number = 1;
-    for (let i = 0; i < 8; i++) {
-      const result: any = (await dbQuery(query, [
-        rollNo,
-        year,
-        sem,
-      ])) as RevalRow[];
-      let subCodes: string[] = [];
-      let subNames: string[] = [];
-      result.forEach((val: { subCode: string; subName: string }) => {
-        subCodes.push(val.subCode);
-        subNames.push(val.subName);
-      });
-      year = i & 1 ? ++year : year;
-      sem = sem == 1 ? 2 : 1;
-      subjDetails.push({ subCodes: subCodes, subNames: subNames });
+    const rollNo: string = req.query.rollNo as string;
+    const exMonth: number = parseInt(req.query.exMonth as string);
+    const exYear: number = parseInt(req.query.exYear as string);
+    if (isAnyUndefined(rollNo, exMonth, exYear)) {
+        res.status(400).json(responses.NotAllParamsGiven);
+        return;
     }
-    subjDetails.forEach((subjDetails, index) => {
-      const semCode = String.fromCharCode("A".charCodeAt(0) + index);
-      subjects[semCode] = subjDetails;
-    });
-    res.json({ subjects, printTableExist: result.length > 0 });
-  } catch (err) {
-    logger.log("error", err);
-    res.json(responses.ErrorWhileDBRequest);
-  }
+    let subjects: Details = {};
+    try {
+        const result = (await dbQuery(
+        `SELECT subCode, subName FROM printReval WHERE rollNo = '${rollNo}';`
+        )) as RevalRow[];
+        // Checking whether the std is already in the print table for registration otherwise fetching std details from the studentInfo table which are not paid
+        const query: string =
+        result.length > 0
+            ? `SELECT subCode, subName FROM printReval WHERE rollNo = ? AND acYear = ? AND sem = ?`
+            : `SELECT std.subCode, std.subName FROM studentInfo std LEFT JOIN paidReEvaluation paidStd ON std.subCode = paidStd.subCode AND std.rollNo = paidStd.rollNo WHERE std.rollNo = ? AND std.exMonth = ${exMonth} AND std.exYear = ${exYear} AND std.acYear = ? AND std.sem = ? AND paidStd.subCode IS NULL AND paidStd.rollNo IS NULL`;
+        let year: number = 1,
+            sem: number = 1,
+            semCode: string = "A";
+        // Fetching Subjects from the table
+        for (let i = 0; i < 8; i++) {
+            const result = (await dbQuery(query, [
+                rollNo,
+                year,
+                sem,
+            ])) as RevalRow[];
+            let subCodes: string[] = [];
+            let subNames: string[] = [];
+            result.forEach((val: { subCode: string; subName: string }) => {
+                subCodes.push(val.subCode);
+                subNames.push(val.subName);
+            });
+            year = i & 1 ? ++year : year;
+            sem = sem == 1 ? 2 : 1;
+            subjects[semCode] =  { subCodes: subCodes, subNames: subNames };
+            semCode = String.fromCharCode(semCode.charCodeAt(0) + 1);
+        }
+        res.json({ subjects, printTableExist: result.length > 0 });
+    } catch (err) {
+        logger.log("error", err);
+        res.json(responses.ErrorWhileDBRequest);
+    }
 }
 
-// Common function for Paid and Register
+// ANCHOR Common function for Print and Register(Paid)
 
 async function revalProcess(req: Request, isPaidTable: boolean) {
     const { body, params } = req;
+    if (isAnyUndefined(body.subjects)) {
+        throw responses.BadRequest;
+    }
     const { rollNo } = params;
-    const { username, subjects, grandTotal, regular } = body.details;
+    const { username, subjects, grandTotal, regular, ip } = body;
     const details = [
         subjects.A,
         subjects.B,
@@ -82,6 +81,7 @@ async function revalProcess(req: Request, isPaidTable: boolean) {
         semCode: string = "A";
     let rows: RevalTable[][] = [];
     const date = dayjs().format("DD-MMM-YY");
+    // Fetching and storing all the details of the subjects to insert them into print/paid tables
     for (const subjects of details) {
         for (let i = 0; i < subjects.subCodes.length; i++) {
             const [subCode, subName] = [subjects.subCodes[i], subjects.subNames[i]];
@@ -95,34 +95,60 @@ async function revalProcess(req: Request, isPaidTable: boolean) {
         await dbQuery(
             `INSERT IGNORE INTO ${tableName} (rollNo, subCode, subName, acYear, sem, regDate, stat, user, grandTotal) VALUES ?`, [rows]
         );
+        if (isPaidTable) await dbQuery(`DELETE FROM printReval WHERE rollNo = '${rollNo}'`);
+        logger.log(`info`, `${req.body.usernameInToken} has added ${rollNo} details in ${tableName} on IP ${ip?.slice(7)}`);
     } catch (err) {
         logger.log("error", err);
         throw responses.ErrorWhileDBRequest;
     }
 }
 
-// Inserting std into printReval Table
+// ANCHOR Inserting std details into printReval Table
 
 export async function printReval(req: Request, res: Response) {
   try {
     await revalProcess(req, false);
   } catch (err) {
     logger.log("error", err);
-    return res.json(responses.ErrorWhileDBRequest);
+    return res.json(err);
   }
   return res.json({ done: true });
 }
 
-// Inserting std into paidReEvaluation table and Deleting std from the printReval table
+// ANCHOR Inserting std details into paidReEvaluation table and Deleting std from the printReval table
 
 export async function registerReval(req: Request, res: Response) {
   const rollNo: string = req.params.rollNo;
   try {
     await revalProcess(req, true);
-    await dbQuery(`DELETE FROM printReval WHERE rollNo = '${rollNo}'`);
   } catch (err) {
     logger.log("error", err);
-    return res.json(responses.ErrorWhileDBRequest);
+    return res.json(err);
   }
   return res.json({ done: true });
+}
+
+export async function deleteFromReval(req: Request, res: Response) {
+    let year = parseInt(req.query.acYear as string) as number;
+    let sem = parseInt(req.query.sem as string) as number;
+    try {
+        if (year === 0 && sem === 0) {
+            await dbQuery("TRUNCATE paidReEvaluation");
+            res.send({ deleted: true });
+            return;
+        }
+        let query: string = "DELETE FROM paidReEvaluation where ";
+        if (year !== 0) {
+            query += `acYear = ${year}`;
+        }
+        if (sem !== 0) {
+            if (year !== 0) query += " and ";
+            query += `sem = ${sem}`;
+        }
+        await dbQuery(query);
+        res.json({ done: "true" });
+    } catch (err) {
+        logger.log("error", err);
+        res.json(responses.ErrorWhileDBRequest);
+    }
 }
